+++ src/main/java/fi/otavanopisto/kuntaapi/server/id/FileId.java
@@ -1,10 +1,7 @@
-package fi.otavanopisto.kuntaapi.server.integrations;
+package fi.otavanopisto.kuntaapi.server.id;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
-
-import fi.otavanopisto.kuntaapi.server.id.Id;
-import fi.otavanopisto.kuntaapi.server.id.IdType;
 
 /**
  * Class representing file id
