+++ src/main/java/fi/otavanopisto/kuntaapi/server/id/IdController.java
@@ -8,7 +8,6 @@ import javax.inject.Inject;
 
 import org.apache.commons.lang3.StringUtils;
 
-import fi.otavanopisto.kuntaapi.server.integrations.FileId;
 import fi.otavanopisto.kuntaapi.server.integrations.KuntaApiConsts;
 
 /**
