+++ src/main/java/fi/otavanopisto/kuntaapi/server/id/IdProvider.java
@@ -1,7 +1,5 @@
 package fi.otavanopisto.kuntaapi.server.id;
 
-import fi.otavanopisto.kuntaapi.server.integrations.FileId;
-
 /**
  * Id translator provider interface
  *
